+++ src/RPKI/UniqueId.hs
@@ -9,4 +9,4 @@ thisExecutableVersion :: ExecutableVersion
 thisExecutableVersion = ExecutableVersion $ rpkiProverVersion <> " " <> 
     -- The content is to be updated by the 'src-hash' script 
     -- that calculates hash of the source tree and configuration/build files     
-    "srcHash#f5352f8c149e757c2b4fc84e25110091c96fd1ac7fac6726ed23530feea39b8a#srcHash"
+    "srcHash#b67f9cc81783c0a77b60d4dc07f4022007ddbbd8c8ecb7a2bc89f2332d4c89a8#srcHash"
