+++ test/src/RPKI/FetchSpec.hs
@@ -14,7 +14,7 @@ import qualified Data.IxSet.Typed                as IxSet
 
 import           RPKI.Domain
 import           RPKI.Repository
-import           RPKI.Fetch
+import           RPKI.Fetch.Fetch
 
 
 uriTaSet :: [(Text, Text)] -> UriTaIxSet
